refactor(apollo-vertex): extract sidebar into separate files
