+++ apps/apollo-vertex/registry/sidebar/sidebar-footer.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import * as React from "react";
+import { cn } from "@/lib/utils";
+
+function SidebarFooter({ className, ...props }: React.ComponentProps<"div">) {
+  return (
+    <div
+      data-slot="sidebar-footer"
+      data-sidebar="footer"
+      className={cn("flex flex-col gap-2 p-2", className)}
+      {...props}
+    />
+  );
+}
+
+export { SidebarFooter };
