+++ apps/apollo-vertex/registry/sidebar/sidebar-menu-sub-item.tsx
@@ -1,0 +1,20 @@
+"use client";
+
+import * as React from "react";
+import { cn } from "@/lib/utils";
+
+function SidebarMenuSubItem({
+  className,
+  ...props
+}: React.ComponentProps<"li">) {
+  return (
+    <li
+      data-slot="sidebar-menu-sub-item"
+      data-sidebar="menu-sub-item"
+      className={cn("group/menu-sub-item relative", className)}
+      {...props}
+    />
+  );
+}
+
+export { SidebarMenuSubItem };
