+++ apps/apollo-vertex/registry/sidebar/sidebar-menu.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import * as React from "react";
+import { cn } from "@/lib/utils";
+
+function SidebarMenu({ className, ...props }: React.ComponentProps<"ul">) {
+  return (
+    <ul
+      data-slot="sidebar-menu"
+      data-sidebar="menu"
+      className={cn("flex w-full min-w-0 flex-col gap-1", className)}
+      {...props}
+    />
+  );
+}
+
+export { SidebarMenu };
