+++ apps/apollo-vertex/registry/sidebar/sidebar-trigger.tsx
@@ -1,0 +1,37 @@
+"use client";
+
+import { PanelLeftIcon } from "lucide-react";
+import * as React from "react";
+import { useTranslation } from "react-i18next";
+import { Button } from "@/components/ui/button";
+import { cn } from "@/lib/utils";
+import { useSidebar } from "./sidebar-provider";
+
+function SidebarTrigger({
+  className,
+  onClick,
+  ...props
+}: React.ComponentProps<typeof Button>) {
+  const { t } = useTranslation();
+  const { toggleSidebar } = useSidebar();
+
+  return (
+    <Button
+      data-sidebar="trigger"
+      data-slot="sidebar-trigger"
+      variant="ghost"
+      size="icon"
+      className={cn("size-7", className)}
+      onClick={(event) => {
+        onClick?.(event);
+        toggleSidebar();
+      }}
+      {...props}
+    >
+      <PanelLeftIcon />
+      <span className="sr-only">{t("toggle_sidebar")}</span>
+    </Button>
+  );
+}
+
+export { SidebarTrigger };
